// src/contexts/Starship.type.ts
export default interface StarshipType {
  name: string;
  model: string;
  manufacturer: string;
  cost_in_credits: string;
  length: string;
  max_atmosphering_speed: string;
  crew: string;
  passengers: string;
  cargo_capacity: string;
  consumables: string;
  hyperdrive_rating: string;
  MGLT: string;
  starship_class: string;
  pilots?: (null)[] | null;
  films?: (string)[] | null;
  created: string;
  edited: string;
  url: string;
  [index: string]: any;
}